requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
asynctest==0.13.0
attrs==23.2.0
aws-lambda-powertools==1.25.5
aws-xray-sdk==2.12.1
backtrader==1.9.78.123
boto3==1.33.13
botocore==1.33.13
certifi==2023.11.17
charset-normalizer==3.3.2
cycler==0.11.0
Django==3.2.23
djangorestframework==3.14.0
fastjsonschema==2.19.1
fonttools==4.38.0
frozenlist==1.3.3
fyers-apiv3==3.0.6
idna==3.6
importlib-metadata==6.7.0
jmespath==0.10.0
kiwisolver==1.4.5
matplotlib==3.5.3
multidict==6.0.4
numpy==1.21.6
packaging==23.2
pandas==1.3.5
Pillow==9.5.0
pyotp==2.9.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.8.2
six==1.16.0
sqlparse==0.4.4
TA-Lib==0.4.28
typing_extensions==4.7.1
urllib3==1.26.18
websocket-client==1.6.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.15.0
python-decouple==3.8.0
